Move job dashboard shortcode methods to their own class
